# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API сервиса аутентификации
  version: 1.0.0
  summary: Аутентификация и вспомогательное API сервиса auth-service
  description: |
    Канонический API-контракт для текущего backend-сервиса.

    Модель аутентификации:
    - Access token возвращается в теле ответа как `accessToken`.
    - Refresh token хранится в HttpOnly cookie `refresh_token`.
    - Идентификатор refresh-сессии хранится в HttpOnly cookie `session_id`.
    - Для cookie-аутентифицированных эндпоинтов требуется CSRF-защита. Клиент должен сначала вызвать
      `GET /api/v1/auth/csrf`, затем отправить полученный токен в заголовке `X-XSRF-TOKEN` вместе
      с cookie `XSRF-TOKEN`.
    - `X-Refresh-Attempt-Id` настоятельно рекомендуется для `/api/v1/auth/refresh`. Повторное
      использование того же значения при сетевом retry позволяет backend безопасно повторить
      результат того же refresh-запроса.

    Политика Origin для браузерных запросов:
    - `allowedOrigins` в настройках auth client является источником истины для CORS.
    - Браузерные запросы с заголовком `Origin` разрешены только если origin принадлежит включенному auth client.
    - Запросы регистрации и входа дополнительно требуют, чтобы `Origin` принадлежал запрошенному `clientId`.
    - Refresh-запросы требуют, чтобы `Origin` принадлежал client, сохраненному в refresh-сессии.
    - Bearer-запросы требуют, чтобы `Origin` принадлежал claim `client_id` в access token.
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/admin/users:
    get:
      tags: [Администрирование]
      summary: Получить список пользователей
      operationId: adminListUsers
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: query
          in: query
          required: false
          description: Поиск по username/email без учета регистра.
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Пользователи возвращены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUserResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/audit-events:
    get:
      tags: [Администрирование]
      summary: Получить список событий аудита
      operationId: adminListAuditEvents
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: eventType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AuditEventType'
        - name: userId
          in: query
          required: false
          description: Ищет события, где пользователь является actor или target.
          schema:
            type: string
            format: uuid
        - name: actorUserId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: targetUserId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: username
          in: query
          required: false
          schema:
            type: string
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
        - name: ip
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Включительная нижняя граница для occurredAt.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Включительная верхняя граница для occurredAt.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: События аудита возвращены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEventResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/auth-clients:
    get:
      tags: [Администрирование]
      summary: Получить список auth clients
      operationId: adminListAuthClients
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список auth clients возвращен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthClientResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    post:
      tags: [Администрирование]
      summary: Создать auth client
      operationId: adminCreateAuthClient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthClientCreateRequest'
      responses:
        '200':
          description: Auth client создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthClientResponse'
        '400':
          description: Ошибка валидации или невалидный origin auth client
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль SUPER_ADMIN
        '409':
          description: Auth client уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/auth-clients/{clientId}:
    get:
      tags: [Администрирование]
      summary: Получить auth client
      operationId: adminGetAuthClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      responses:
        '200':
          description: Auth client возвращен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthClientResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '404':
          description: Auth client не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    put:
      tags: [Администрирование]
      summary: Обновить auth client
      operationId: adminUpdateAuthClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthClientUpdateRequest'
      responses:
        '200':
          description: Auth client обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthClientResponse'
        '400':
          description: Ошибка валидации или невалидный origin auth client
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль SUPER_ADMIN
        '404':
          description: Auth client не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/auth-clients/{clientId}/enable:
    post:
      tags: [Администрирование]
      summary: Включить auth client
      operationId: adminEnableAuthClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      responses:
        '200':
          description: Auth client включен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthClientResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль SUPER_ADMIN
        '404':
          description: Auth client не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/auth-clients/{clientId}/disable:
    post:
      tags: [Администрирование]
      summary: Отключить auth client
      operationId: adminDisableAuthClient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      responses:
        '200':
          description: Auth client отключен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthClientResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль SUPER_ADMIN
        '404':
          description: Auth client не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/users/{userId}:
    get:
      tags: [Администрирование]
      summary: Получить пользователя
      operationId: adminGetUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Пользователь возвращен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '404':
          description: Пользователь не найден
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/users/{userId}/sessions:
    get:
      tags: [Администрирование]
      summary: Получить список сессий пользователя
      operationId: adminListUserSessions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SessionStatus'
      responses:
        '200':
          description: Сессии возвращены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '404':
          description: Пользователь не найден
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/users/{userId}/logout-all:
    post:
      tags: [Администрирование]
      summary: Отозвать все сессии пользователя
      operationId: adminLogoutUserEverywhere
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          description: Сессии пользователя отозваны
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '404':
          description: Пользователь не найден
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/users/{userId}/sessions/{sessionId}/revoke:
    post:
      tags: [Администрирование]
      summary: Отозвать сессию пользователя
      operationId: adminRevokeUserSession
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Сессия пользователя отозвана
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль ADMIN или SUPER_ADMIN
        '404':
          description: Пользователь не найден
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/users/{userId}/ban:
    put:
      tags: [Администрирование]
      summary: Заблокировать пользователя
      description: |
        Создает активный бан и отзывает все сессии пользователя.
        `expiresAt = null` означает постоянный бан.
        ADMIN может блокировать только обычных пользователей.
        SUPER_ADMIN может блокировать привилегированных пользователей, кроме себя и защищенного bootstrap.
      operationId: adminBanUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBanRequest'
      responses:
        '200':
          description: Пользователь заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          description: Некорректная причина или срок бана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Недостаточно прав, self-ban или пользователь защищен от бана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Пользователь не найден
        '409':
          description: Пользователь уже имеет активный бан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    delete:
      tags: [Администрирование]
      summary: Снять активный бан пользователя
      description: Повторный вызов для пользователя без активного бана является успешным.
      operationId: adminUnbanUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Активный бан снят или отсутствовал
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Недостаточно прав или попытка управлять собственным баном
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Пользователь не найден
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/admin/users/{userId}/roles/admin:
    put:
      tags: [Администрирование]
      summary: Выдать роль ADMIN
      operationId: adminGrantAdminRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          description: Роль ADMIN выдана
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль SUPER_ADMIN
        '404':
          description: Пользователь не найден
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    delete:
      tags: [Администрирование]
      summary: Отозвать роль ADMIN
      operationId: adminRevokeAdminRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          description: Роль ADMIN отозвана
        '401':
          description: Bearer token отсутствует или невалиден
        '403':
          description: Требуется роль SUPER_ADMIN
        '404':
          description: Пользователь не найден
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
components:
  responses:
    RateLimitExceeded:
      description: Превышен rate limit для текущего scope запроса.
      headers:
        Retry-After:
          description: Рекомендуемая задержка перед повторной попыткой, в секундах.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: RATE_LIMIT_EXCEEDED
            message: Too many requests. Please try again later
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ClientIdPath:
      name: clientId
      in: path
      required: true
      schema:
        type: string
        maxLength: 128
...
  schemas:
...
    SessionStatus:
      type: string
      enum:
        - ACTIVE
        - REVOKED
        - EXPIRED
        - COMPROMISED
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [timestamp, status, error, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        path:
          type: string
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - INVALID_REFRESH_TOKEN
            - REFRESH_TOKEN_EXPIRED
            - REFRESH_TOKEN_REUSE_DETECTED
            - INVALID_CREDENTIALS
            - USERNAME_ALREADY_TAKEN
            - EMAIL_ALREADY_REGISTERED
            - REFRESH_TOKEN_ALREADY_PROCESSED
            - INVALID_CLIENT
            - USER_NOT_FOUND
            - AUTH_CLIENT_ALREADY_EXISTS
            - AUTH_CLIENT_NOT_FOUND
            - RATE_LIMIT_EXCEEDED
            - ORIGIN_NOT_ALLOWED
            - INVALID_AUTH_CLIENT_ORIGIN
            - REQUEST_BODY_TOO_LARGE
            - USER_BANNED
            - INVALID_USER_BAN
            - USER_BAN_FORBIDDEN
            - USER_ALREADY_BANNED
        message:
          type: string
    AdminUserResponse:
      type: object
      additionalProperties: false
      required: [id, username, email, createdAt, banProtected, activeBan, roles]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        banProtected:
          type: boolean
        activeBan:
          oneOf:
            - $ref: '#/components/schemas/UserBanResponse'
            - type: 'null'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
...
    AuditEventResponse:
      type: object
      additionalProperties: false
      required: [id, eventType, occurredAt, details]
      properties:
        id:
          type: string
          format: uuid
        eventType:
          $ref: '#/components/schemas/AuditEventType'
        occurredAt:
          type: string
          format: date-time
        actorUserId:
          type: string
          format: uuid
        targetUserId:
          type: string
          format: uuid
        username:
          type: string
        sessionId:
          type: string
        ip:
          type: string
        userAgent:
          type: string
        details:
          type: object
          additionalProperties: true
    AuditEventType:
      type: string
      enum:
        - USER_REGISTERED
        - USER_LOGGED_IN
        - USER_LOGIN_FAILED
        - USER_LOGGED_OUT
        - USER_LOGGED_OUT_ALL
        - SESSION_REVOKED
        - ALL_SESSIONS_REVOKED
        - REFRESH_TOKEN_ROTATED
        - REFRESH_TOKEN_REUSE_DETECTED
        - ADMIN_ROLE_GRANTED
        - ADMIN_ROLE_REVOKED
        - USER_BANNED
        - USER_UNBANNED
        - SUPER_ADMIN_BOOTSTRAPPED
        - RATE_LIMIT_EXCEEDED
        - AUTH_CLIENT_CREATED
        - AUTH_CLIENT_UPDATED
        - AUTH_CLIENT_ENABLED
        - AUTH_CLIENT_DISABLED
    AuthClientCreateRequest:
      type: object
      additionalProperties: false
      required: [clientId, name, accessTokenTtlSeconds, refreshTokenTtlSeconds, tokenAudience, allowedOrigins]
      properties:
        clientId:
          type: string
          maxLength: 128
          minLength: 1
          example: budget-manager-web
        name:
          type: string
          maxLength: 128
          minLength: 1
          example: Budget Manager Web
        enabled:
          type: boolean
          default: true
        accessTokenTtlSeconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 86400
          description: TTL access token для конкретного auth client в секундах. Фактический TTL ограничен `APP_SECURITY_ACCESS_TOKEN_TTL`.
          example: 600
        refreshTokenTtlSeconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 7776000
          example: 1209600
        tokenAudience:
          type: string
          maxLength: 128
          minLength: 1
          example: budget-manager
        allowedOrigins:
          type: array
          maxItems: 50
          description: |
            Браузерные origins, разрешенные для этого auth client. Списком управляют пользователи с ролью SUPER_ADMIN.
            Значения используются для CORS preflight и привязки браузерных запросов к найденному auth client.
            Пример формата origin: `https://app.example.com`; путь указывать не нужно.
          items:
            type: string
            minLength: 1
            maxLength: 256
          example:
            - http://localhost:3000
            - http://localhost:5173
    AuthClientUpdateRequest:
      type: object
      additionalProperties: false
      required: [name, accessTokenTtlSeconds, refreshTokenTtlSeconds, tokenAudience, allowedOrigins]
      properties:
        name:
          type: string
          maxLength: 128
          minLength: 1
          example: Budget Manager Web
        accessTokenTtlSeconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 86400
          description: TTL access token для конкретного auth client в секундах. Фактический TTL ограничен `APP_SECURITY_ACCESS_TOKEN_TTL`.
          example: 600
        refreshTokenTtlSeconds:
          type: integer
          format: int64
          minimum: 1
          maximum: 7776000
          example: 1209600
        tokenAudience:
          type: string
          maxLength: 128
          minLength: 1
          example: budget-manager
        allowedOrigins:
          type: array
          maxItems: 50
          description: |
            Полный список браузерных origins, разрешенных для этого auth client.
            Значения используются для CORS preflight и привязки браузерных запросов к найденному auth client.
            Пример формата origin: `https://app.example.com`; путь указывать не нужно.
          items:
            type: string
            minLength: 1
            maxLength: 256
          example:
            - http://localhost:3000
            - http://localhost:5173
    AuthClientResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - clientId
        - name
        - enabled
        - accessTokenTtlSeconds
        - refreshTokenTtlSeconds
        - tokenAudience
        - allowedOrigins
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        clientId:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        accessTokenTtlSeconds:
          type: integer
          format: int64
        refreshTokenTtlSeconds:
          type: integer
          format: int64
        tokenAudience:
          type: string
        allowedOrigins:
          type: array
          description: Браузерные origins, разрешенные для CORS и привязки client-origin.
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Role:
      type: string
      enum:
        - USER
        - ADMIN
        - SUPER_ADMIN